+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+- Add support for object(`SpanOptions`) as an argument for `startChildSpan` function, similar to `startRootSpan`.
+
 ## 0.0.2 - 2019-04-29
 
 Fix: add JS bundles and source maps to the NPM files for @opencensus/web-all